- moved changeNoteToPublic to domain/entities/notePublic.ts - Got rid of public noteSetting
