--- src/domain/entities/noteSettingsPublic.ts
@@ -1,14 +1,0 @@
-/**
- * Notes settings entity
- */
-export default interface NoteSettingsPublic {
-  /**
-   * Custom hostname
-   */
-  customHostname?: string ;
-
-  /**
-   * Is note public for everyone or only for collaborators
-   */
-  isPublic: boolean;
-}
